✅(ingestion-infrastructure) make load documents integration container independant
