+++ src/tycho/tests/ingestion/integration/test_integration_load_documents.py
@@ -1,76 +1,65 @@
+import re
+from datetime import datetime, timezone
+from typing import Optional
+from unittest.mock import patch
+
 import pytest
 from asgiref.sync import sync_to_async
-from django.contrib.auth.models import User
-from django.core.files.uploadedfile import SimpleUploadedFile
-from django.urls import reverse
-from faker import Faker
-from rest_framework import status
-from rest_framework.test import APIClient
-from rest_framework_simplejwt.tokens import RefreshToken
 
 from application.ingestion.interfaces.load_documents_input import LoadDocumentsInput
 from application.ingestion.interfaces.load_operation_type import LoadOperationType
-from domain.entities.document import DocumentType
+from application.ingestion.usecases import load_offers
+from config.app_config import AppConfig
+from domain.entities.document import Document, DocumentType
+from domain.exceptions.document_error import InvalidDocumentTypeError
 from infrastructure.di.ingestion.ingestion_container import IngestionContainer
+from infrastructure.di.shared.shared_container import SharedContainer
 from infrastructure.django_apps.ingestion.models.raw_document import RawDocument
 from infrastructure.gateways.shared.logger import LoggerService
-from infrastructure.repositories.ingestion.postgres_document_repository import (
-    PostgresDocumentRepository,
-)
-from infrastructure.repositories.shared.postgres_concours_repository import (
-    PostgresConcoursRepository,
-)
-from infrastructure.repositories.shared.postgres_corps_repository import (
-    PostgresCorpsRepository,
-)
-from infrastructure.repositories.shared.postgres_offers_repository import (
-    PostgresOffersRepository,
-)
 from tests.factories.ingres_corps_factories import IngresCorpsApiResponseFactory
-from tests.fixtures.fixture_loader import load_fixture
-from tests.utils.in_memory_vector_repository import InMemoryVectorRepository
-from tests.utils.mock_embedding_generator import MockEmbeddingGenerator
+from tests.factories.talentsoft_factories import TalentsoftDetailOfferFactory
+from tests.ingestion.unit.external_gateways.utils import (
+    cached_token,
+    offers_response,
+)
 
-fake = Faker()
+PORT = 6333
+MAX_ITERATIONS = 3
+
+
+@pytest.fixture
+def documents_ingestion_container():
+    container = IngestionContainer()
+
+    shared_container = SharedContainer()
+
+    app_config = AppConfig.from_django_settings()
+    shared_container.app_config.override(app_config)
+
+    logger_service = LoggerService()
+    shared_container.logger_service.override(logger_service)
+
+    container.shared_container.override(shared_container)
+
+    container.app_config.override(app_config)
+    container.logger_service.override(logger_service)
+
+    return container
+
+
+@pytest.fixture
+def test_app_config(documents_ingestion_container):
+    return documents_ingestion_container.app_config()
+
+
+@pytest.fixture
+def load_documents_usecase(documents_ingestion_container):
+    return documents_ingestion_container.load_documents_usecase()
 
 
 class TestIntegrationCorpsLoadDocumentsUseCase:
-    @pytest.fixture(name="documents_integration_usecase")
-    def documents_integration_usecase_fixture(self, test_app_config):
-        container = IngestionContainer()
-        logger_service = LoggerService()
-
-        # Configuration de base
-        container.logger_service.override(logger_service)
-        container.app_config.override(test_app_config)
-        container.shared_container.app_config.override(test_app_config)
-
-        # Mock embedding generator
-        embedding_fixtures = load_fixture("embedding_fixtures.json")
-        embedding_generator = MockEmbeddingGenerator(embedding_fixtures)
-        container.shared_container.embedding_generator.override(embedding_generator)
-
-        # Repositories Postgres (vrais pour l'intégration)
-        postgres_document_repo = PostgresDocumentRepository()
-        container.document_repository.override(postgres_document_repo)
-
-        postgres_corps_repo = PostgresCorpsRepository(logger_service)
-        container.shared_container.corps_repository.override(postgres_corps_repo)
-
-        postgres_concours_repo = PostgresConcoursRepository(logger_service)
-        container.shared_container.concours_repository.override(postgres_concours_repo)
-
-        postgres_offers_repo = PostgresOffersRepository(logger_service)
-        container.shared_container.offers_repository.override(postgres_offers_repo)
-
-        # InMemoryVectorRepository pour éviter Qdrant
-        in_memory_vector_repo = InMemoryVectorRepository(logger_service)
-        container.shared_container.vector_repository.override(in_memory_vector_repo)
-
-        return container.load_documents_usecase()
-
     async def test_execute_returns_zero_when_no_documents(
-        self, db, documents_integration_usecase, test_app_config, httpx_mock
+        self, db, load_documents_usecase, test_app_config, httpx_mock
     ):
         # Mock OAuth token endpoint
         httpx_mock.add_response(
@@ -93,12 +82,12 @@ class TestIntegrationCorpsLoadDocumentsUseCase:
             operation_type=LoadOperationType.FETCH_FROM_API,
             kwargs={"document_type": DocumentType.CORPS},
         )
-        result = await documents_integration_usecase.execute(input_data)
+        result = await load_documents_usecase.execute(input_data)
         assert result["created"] == 0
         assert result["updated"] == 0
 
     async def test_execute_returns_correct_count_with_documents(
-        self, db, documents_integration_usecase, test_app_config, httpx_mock
+        self, db, load_documents_usecase, test_app_config, httpx_mock
     ):
         api_response = IngresCorpsApiResponseFactory.build()
         api_data = [doc.model_dump(mode="json") for doc in api_response.documents]
@@ -124,7 +113,7 @@ class TestIntegrationCorpsLoadDocumentsUseCase:
             operation_type=LoadOperationType.FETCH_FROM_API,
             kwargs={"document_type": DocumentType.CORPS},
         )
-        result = await documents_integration_usecase.execute(input_data)
+        result = await load_documents_usecase.execute(input_data)
         assert result["created"] == len(api_data)
         assert result["updated"] == 0
 
@@ -139,95 +128,129 @@ class TestIntegrationCorpsLoadDocumentsUseCase:
         assert saved_count == len(api_data)
 
 
-@pytest.fixture(name="api_client")
-def api_client_fixture():
-    return APIClient()
+@pytest.fixture
+def load_offers_usecase(documents_ingestion_container):
+    return documents_ingestion_container.load_offers_usecase()
 
 
-@pytest.fixture(name="user")
-def user_fixture(db):
-    return User.objects.create_user(
-        username=fake.name(), email=fake.email(), password=fake.password()
+def format_modification_date(dt: datetime) -> str:
+    return dt.strftime("%Y-%m-%dT%H:%M:%S.") + f"{dt.microsecond // 1000:03d}"
+
+
+def offersummaries_url(client, start: int) -> re.Pattern:
+    return re.compile(
+        rf"{re.escape(str(client.base_url))}/api/v2/offersummaries\?count=100&start={start}"
     )
 
 
-@pytest.fixture(name="authenticated_client")
-def authenticated_client_fixture(api_client, user):
-    refresh = RefreshToken.for_user(user)
-    token = str(refresh.access_token)
-    api_client.credentials(HTTP_AUTHORIZATION=f"Bearer {token}")
-    return api_client
-
-
-@pytest.fixture(name="valid_csv_content")
-def valid_csv_content_fixture():
-    return (
-        "N° NOR;Ministère;Catégorie;Corps;"
-        "Grade;Année de référence;Nb postes total\n"
-        "INTB2400001C;Ministère de l'Intérieur;A;Attaché;Attaché;2024;10\n"
-        "INTB2400002C;Ministère de l'Intérieur;B;Secrétaire;Secrétaire;2024;5\n"
+def getoffer_url(client, reference: str = ".*") -> re.Pattern:
+    return re.compile(
+        rf"{re.escape(str(client.base_url))}/api/v2/offers/getoffer\?reference={reference}&sort=modificationDate"
     )
 
 
-@pytest.fixture(name="invalid_csv_content")
-def invalid_csv_content_fixture():
-    return (
-        "N° NOR;Ministère;Catégorie;Corps;Grade;"
-        "Année de référence;Nb postes total\n"
-        ";Ministère de l'Intérieur;A;Attaché;Attaché;2024;10\n"
-        "INTB2400002C;;B;Secrétaire;Secrétaire;2024;5\n"
+def mock_offersummaries_response(
+    client,
+    httpx_mock,
+    *,
+    start: int = 1,
+    count_offers: int,
+    has_more: bool = False,
+    offers: Optional[list] = None,
+):
+    json = (
+        offers_response(count=count_offers, has_more=has_more, offers=offers)
+        if offers
+        else offers_response(count=count_offers, has_more=has_more)
+    )
+    httpx_mock.add_response(
+        method="GET", url=offersummaries_url(client, start), json=json, status_code=200
     )
 
 
-def make_csv_file(content, filename="test.csv"):
-    return SimpleUploadedFile(
-        filename, content.encode("utf-8"), content_type="text/csv"
+def mock_detail_offer_response(client, httpx_mock, *, offer: Optional[dict] = None):
+    reference = offer["reference"] if offer else None
+    json = TalentsoftDetailOfferFactory.build(
+        **(
+            {"reference": reference, "modificationDate": offer["modificationDate"]}
+            if offer
+            else {}
+        )
+    ).model_dump()
+    httpx_mock.add_response(
+        method="GET",
+        url=getoffer_url(client, reference or ".*"),
+        json=json,
+        status_code=200,
+        is_reusable=offer is None,
     )
 
 
-class TestConcoursUploadView:
-    url = reverse("ingestion:concours_upload")
+def make_document(offer: dict) -> Document:
+    reference = offer["reference"]
+    versant_code_obj = offer["salaryRange"]
+    versant = versant_code_obj["clientCode"] if versant_code_obj else "UNK"
+    return Document(
+        external_id=f"{versant}-{reference}",
+        raw_data=offer,
+        type=DocumentType.OFFERS,
+        created_at=datetime.now(timezone.utc),
+    )
 
-    def test_unauthenticated_access(self, api_client, valid_csv_content):
-        response = api_client.post(
-            self.url, {"file": make_csv_file(valid_csv_content)}, format="multipart"
-        )
-        assert response.status_code == status.HTTP_401_UNAUTHORIZED
 
-    def test_no_file_provided(self, authenticated_client):
-        response = authenticated_client.post(self.url, {}, format="multipart")
-        assert response.status_code == status.HTTP_400_BAD_REQUEST
-        assert response.data["error"] == "No file provided"
+class TestIntegrationLoadOffersUseCase:
+    async def test_raises_for_non_offers_type(self, db, load_offers_usecase):
+        input_data = LoadDocumentsInput(
+            operation_type=LoadOperationType.FETCH_FROM_API,
+            kwargs={"document_type": DocumentType.CORPS},
+        )
 
-    def test_invalid_file_format(self, authenticated_client):
-        txt_file = SimpleUploadedFile(
-            "test.txt", b"not a csv file", content_type="text/plain"
-        )
-        response = authenticated_client.post(
-            self.url, {"file": txt_file}, format="multipart"
-        )
-        assert response.status_code == status.HTTP_400_BAD_REQUEST
-        assert response.data["error"] == "File must be a CSV"
+        with pytest.raises(InvalidDocumentTypeError):
+            await load_offers_usecase.execute(input_data)
 
-    def test_validation_errors(self, authenticated_client, invalid_csv_content):
-        response = authenticated_client.post(
-            self.url, {"file": make_csv_file(invalid_csv_content)}, format="multipart"
-        )
-        assert response.status_code == status.HTTP_400_BAD_REQUEST
-        assert response.data["error"] == "No valid rows found"
-        assert "validation_errors" in response.data
-        assert len(response.data["validation_errors"]) == 2  # noqa
+    @pytest.mark.httpx_mock
+    async def test_stops_when_fetch_by_type_returns_no_documents(
+        self, db, httpx_mock, documents_ingestion_container, load_offers_usecase
+    ):
+        client = documents_ingestion_container.talentsoft_front_client()
+        client.cached_token = cached_token()
 
-    def test_success_response(self, db, authenticated_client, valid_csv_content):
-        response = authenticated_client.post(
-            self.url, {"file": make_csv_file(valid_csv_content)}, format="multipart"
+        mock_offersummaries_response(client, httpx_mock, count_offers=0)
+
+        input_data = LoadDocumentsInput(
+            operation_type=LoadOperationType.FETCH_FROM_API,
+            kwargs={"document_type": DocumentType.OFFERS},
         )
-        assert response.status_code == status.HTTP_201_CREATED
-        assert response.data["status"] == "success"
-        assert response.data["total_rows"] == 2  # noqa
-        assert response.data["valid_rows"] == 2  # noqa
-        assert response.data["invalid_rows"] == 0
-        assert (
-            "Successfully processed 2 valid concours records"
-            in response.data["message"]
+
+        with patch.object(
+            load_offers_usecase.document_repository,
+            "find_by_external_ids",
+            wraps=load_offers_usecase.document_repository.find_by_external_ids,
+        ) as repository_find_by_external_ids:
+            result = await load_offers_usecase.execute(input_data)
+
+        repository_find_by_external_ids.assert_not_called()
+        assert result == {"created": 0, "updated": 0, "errors": []}
+
+    async def test_stops_after_max_iterations(
+        self, db, httpx_mock, documents_ingestion_container, load_offers_usecase
+    ):
+        client = documents_ingestion_container.talentsoft_front_client()
+        client.cached_token = cached_token()
+
+        for start in range(1, MAX_ITERATIONS):
+            mock_offersummaries_response(
+                client, httpx_mock, start=start, count_offers=1, has_more=True
+            )
+        mock_detail_offer_response(client, httpx_mock)
+
+        input_data = LoadDocumentsInput(
+            operation_type=LoadOperationType.FETCH_FROM_API,
+            kwargs={"document_type": DocumentType.OFFERS},
         )
+
+        with patch.object(load_offers, "MAX_ITERATIONS", MAX_ITERATIONS):
+            await load_offers_usecase.execute(input_data)
+
+        # (MAX_ITERATIONS - 1) * (getsummaries + getoffer calls)
+        assert len(httpx_mock.get_requests()) == (MAX_ITERATIONS - 1) * 2
